Change window dimensions to integers and add system images
Converted window dimension properties from `double` to `int` for consistency and simplified type handling. Added new system images to support additional console platforms.

+++ SimpleLauncher/MainWindow.xaml.cs
@@ -265,10 +265,10 @@ public partial class MainWindow : INotifyPropertyChanged
 
     private void SaveApplicationSettings()
     {
-        _settings.MainWindowWidth = Width;
-        _settings.MainWindowHeight = Height;
-        _settings.MainWindowTop = Top;
-        _settings.MainWindowLeft = Left;
+        _settings.MainWindowWidth = (int)Width;
+        _settings.MainWindowHeight = (int)Height;
+        _settings.MainWindowTop = (int)Top;
+        _settings.MainWindowLeft = (int)Left;
         _settings.MainWindowState = WindowState.ToString();
 
         // Set other settings from the application's current state
